Make post its own component
