+++ client/pages/home.jsx
@@ -1,10 +1,7 @@
 import React from 'react';
 import Navbar from '../components/navbar';
-import RunInfo from '../components/runInfo';
-import { Container, Carousel, Modal, Button } from 'react-bootstrap';
-import formatDistanceStrict from 'date-fns/formatDistanceStrict';
-import DropdownCustom from '../components/dropdown';
-import Like from '../components/like';
+import { Modal, Button } from 'react-bootstrap';
+import Post from '../components/post';
 
 export default class Home extends React.Component {
   constructor(props) {
@@ -69,71 +66,7 @@ export default class Home extends React.Component {
   render() {
     if (!this.state.posts) return;
 
-    const options = [
-      { href: '#edit', text: 'Edit' },
-      { href: '#home', text: 'Remove' }
-    ];
-
-    let imgSrc, postData;
-
-    const posts = this.state.posts.map((post, index) => {
-      postData = post;
-      let carouselItems = post.images.map(image => {
-        if (image.on) {
-          imgSrc = image.url;
-          return (
-            <Carousel.Item key={image.url} id={image.url} className="text-light position-relative">
-              <img src={image.url} />
-            </Carousel.Item>
-          );
-        } else {
-          return null;
-        }
-      });
-
-      carouselItems = carouselItems.filter(x => x !== null);
-
-      const dropdown = (postData.accountId === this.props.login.user.accountId)
-        ? (
-          <DropdownCustom direction="horizontal" options={options} saveRunId={this.props.saveRunId} savePostId={this.props.savePostId} showModal={this.showModal} />
-          )
-        : (
-            null
-          );
-
-      const carousel = (carouselItems.length !== 1)
-        ? <Carousel interval={null}>{carouselItems}</Carousel>
-        : <div className='single-image-post'><img src={imgSrc} /></div>;
-
-      const now = new Date();
-      const then = new Date(postData.postedAt);
-      const result = formatDistanceStrict(then, now, { includeSeconds: true, addSuffix: true });
-
-      return (
-        <Container className="outer" key={postData.postId} runid={postData.runId} postid={postData.postId}>
-          <div className='d-flex'>
-            <p className='desktop-username m-2 mb-1 ps-3'>{postData.username}</p>
-            <div className='ms-auto dropdown-ellipsis align-self-center me-3'>
-              {dropdown}
-            </div>
-          </div>
-          {carousel}
-          <RunInfo key={postData.postId} postData={postData} />
-          <Container className='mt-1 mb-3'>
-            <div className='icons'>
-              <Like login={this.props.login} />
-            </div>
-            <div className='d-flex'>
-              <p className='small mb-0'>{postData.username}</p>
-              <p className='small ps-2 mb-0'>{postData.caption}</p>
-            </div>
-            <div>
-              <p className='small mb-0'>{result}</p>
-            </div>
-          </Container>
-        </Container>
-      );
-    });
+    const posts = this.state.posts.map(post => <Post key={post.postId} postData={post} login={this.props.login} />);
 
     return (
       <div className="home-page" >
